Virtual Whiteboard
This repo contains script files for a Virtual Whiteboard Application.
The Application was developed as a part of summer project in RoboVITics club. This project is targeted towards those who prefer written or drawn mode of presentation. It can be used by anyone ranging from artists and teachers to business professionals. 
The application utilizes mediapipe and opencv to track hand position and shape and uses the finger as pointer for marking and drawing